- Do ChunkSize compatible with NimbleStreamer which doesn't respect RTMP standard (max chunkSize = 0x003EFFEE) - Fix a RTMP chunkSize reception issue (chunkSize is UInt32 and not UInt16 type) - Other minor changes

+++ MonaCore/include/Mona/RTMP/RTMPSession.h
@@ -50,7 +50,7 @@ private:
 	FlashStream::OnStart::Type		onStreamStop;
 
 	UInt8							_handshaking;
-	UInt16							_chunkSize;
+	UInt32							_chunkSize;
 	UInt32							_winAckSize;
 	UInt32							_unackBytes;
 	UInt32							_readBytes;
